requirements.txt
pandas
streamlit
nltk
pillow
sklearn
plotly